\name{internal}
\alias{.lec.CheckSeed}
\alias{.lec.DoCreateStream}
\alias{.lec.GetStateList}
\alias{.lec.Getuniform}
\alias{.lec.SaveStateList}
\alias{.lec.which.stream}

\title{Internal functions}
\description{Internal functions, it is not recommended to use them from outside of the package.
}


\keyword{internal}

