\name{WriteState}
\alias{.lec.WriteState}
\alias{.lec.WriteStateFull}
\title{Output of the current state of streams}
\description{
\code{.lec.WriteState} writes the current state of given streams (Cg values).

\code{.lec.WriteStateFull} writes the values of all internal state variables
                        of given streams. 
}
\usage{
.lec.WriteState (names)

.lec.WriteStateFull (names)
}
\arguments{
  \item{names}{a character string or a vector of character strings
    naming the streams.}
}
\value{None}
\keyword{distribution}
